+++ hls_ip/fractal/fractal.h
@@ -8,9 +8,9 @@
 
 #include <cstdint>
 
-static constexpr std::uint32_t MAX_WIDTH      = 1920u;
-static constexpr std::uint32_t MAX_HEIGHT     = 1080u;
-static constexpr std::uint32_t MAX_ITERATIONS = 255u;
+static constexpr std::uint32_t MAX_WIDTH     = 1920u;
+static constexpr std::uint32_t MAX_HEIGHT    = 1080u;
+static constexpr std::uint8_t MAX_ITERATIONS = 255u;
 
 using video_type  = ap_axiu<24, 1, 1, 1>;
 using stream_type = hls::stream<video_type>;
